On OS(DARWIN), enable ExecutableAllocator to use a constexpr page size instead of calling WTF::pageSize() repeatedly. https://bugs.webkit.org/show_bug.cgi?id=242971
Reviewed by Yusuke Suzuki.

1. On OS(DARWIN), the page size is fixed.  We already rely on this in MarkedBlock.
   Allow ExecutableAllocator to do the same and use the constexpr CeilingOnPageSize value as the
   page size instead of repeatedly calling WTF::pageSize().
2. Moved isPowerOfTwo() to StdLibExtra.h, and made it constexpr.
3. Made roundUpToMultipleOf() constexpr.

   The above working together allows many page size calculations to be constant folded.

4. Added a comment to clarify why each allocatable region in the RegionAllocators must be able to
   fit 2 jump islands in a 128M span (for ARM64), and not just one.  This is because the jumps in
   the jump islands can chain using relative jumps to adjacent islands as well.  Hence, any
   adjacent neighbor jump island needs to be within the ARM64 128M near jump range.

5. Removed some unused code.

* Source/JavaScriptCore/dfg/DFGSpeculativeJIT64.cpp:
(JSC::DFG::SpeculativeJIT::emitCall):
* Source/JavaScriptCore/jit/ExecutableAllocator.cpp:
(JSC::initializeJITPageReservation):
* Source/JavaScriptCore/jit/ExecutableAllocator.h:
* Source/JavaScriptCore/jit/JITCall.cpp:
(JSC::JIT::compileSetupFrame):
* Source/WTF/wtf/PageBlock.cpp:
(WTF::pageSize):
(WTF::pageMask): Deleted.
* Source/WTF/wtf/PageBlock.h:
(WTF::isPageAligned):
(WTF::isPowerOfTwo): Deleted.
* Source/WTF/wtf/StdLibExtras.h:
(WTF::isPowerOfTwo):
(WTF::roundUpToMultipleOfImpl):
(WTF::roundUpToMultipleOf):

Canonical link: https://commits.webkit.org/252775@main

+++ Source/JavaScriptCore/dfg/DFGSpeculativeJIT64.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2011-2021 Apple Inc. All rights reserved.
+ * Copyright (C) 2011-2022 Apple Inc. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -792,7 +792,7 @@ void SpeculativeJIT::emitCall(Node* node)
             loadArgumentsGPR(GPRInfo::returnValueGPR);
             m_jit.move(TrustedImm32(numUsedStackSlots), scratchGPR1);
             emitSetVarargsFrame(m_jit, GPRInfo::returnValueGPR, false, scratchGPR1, scratchGPR1);
-            m_jit.addPtr(TrustedImm32(-(sizeof(CallerFrameAndPC) + WTF::roundUpToMultipleOf(stackAlignmentBytes(), 5 * sizeof(void*)))), scratchGPR1, JITCompiler::stackPointerRegister);
+            m_jit.addPtr(TrustedImm32(-static_cast<int32_t>(sizeof(CallerFrameAndPC) + WTF::roundUpToMultipleOf(stackAlignmentBytes(), 5 * sizeof(void*)))), scratchGPR1, JITCompiler::stackPointerRegister);
             
             callOperation(operationSetupVarargsFrame, GPRInfo::returnValueGPR, JITCompiler::LinkableConstant(m_jit, m_graph.globalObjectFor(node->origin.semantic)), scratchGPR1, argumentsGPR, data->firstVarArgOffset, GPRInfo::returnValueGPR);
             m_jit.exceptionCheck();
